refactor/wallet-navbar
