+++ apps/web-app/src/lib/utils.ts
@@ -4,3 +4,8 @@ import { twMerge } from "tailwind-merge";
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
+
+export function truncateAddress(value: string, start = 4, end = 4): string {
+  if (value.length <= start + end) return value;
+  return `${value.slice(0, start)}…${value.slice(-end)}`;
+}
